refactor(card-builder): modularize editor components
