+++ packages/card-builder/src/components/Palette.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useDraggable } from "@dnd-kit/core";
 import { CSS } from "@dnd-kit/utilities";
-import { ElementDefinition } from "./elements";
+import { ElementDefinition } from "../elements";
 
 function PaletteItem({ def }: { def: ElementDefinition }) {
   const { attributes, listeners, setNodeRef, transform } = useDraggable({
